【修改描述】故障联调过程中BUG修复 【DTS/CR/AR单号】#184 #185 #186 【合入版本】Open Computing Kit for Kunpeng 24.0.0
Match-id-0de5435d9b1b8e6c00da24d2300e964e0bcb02cb

+++ src/flow/flow_type.h
@@ -9,7 +9,7 @@
 
 namespace ock {
 namespace bio {
-enum FlowType : uint8_t {
+enum FlowType : int32_t {
     FLOW_MEMORY = 1,
     FLOW_DISK = 2,
 };
